Succesful Editor build on x64

+++ Mini_Maniacs/mesh.cpp
@@ -31,7 +31,7 @@ void mesh::Write(std::ofstream* s) const
   for (auto& a : verticies) 
   {
     *s << a.position.x << " " << a.position.y << "\n";
-    *s << a.color.r << " " << a.color.g << " " << a.color.b << " " << a.color.a << "\n";
+    *s << static_cast<int>(a.color.r) << " " << static_cast<int>(a.color.g) << " " << static_cast<int>(a.color.b) << " " << static_cast<int>(a.color.a) << "\n";
     *s << a.tex_coord.x << " " << a.tex_coord.y << "\n";
   }
